Remove ignored file

+++ malaria-ios/ItemExtension.swift
@@ -1,11 +1,11 @@
 import Foundation
 
 extension Item{
-    func add(number: Int){
+    func add(number: Int64){
         self.number += number
     }
     
-    func remove(number: Int){
+    func remove(number: Int64){
         self.number = max(self.number - number, 0)
     }
 }
