+++ malaria-ios/Trip.swift
@@ -4,7 +4,7 @@ import CoreData
 class Trip: NSManagedObject {
 
     @NSManaged var medicine: String
-    @NSManaged var cashToBring: Int
+    @NSManaged var cashToBring: Int64
     @NSManaged var reminderDate: NSDate
     @NSManaged var location: String
     @NSManaged var items: NSMutableSet
